Fix GCC7.2 constexpr compilation issue.

+++ src/util/math.cpp
@@ -68,7 +68,7 @@ static constexpr NBitFloatInfo Float16Info =
     -((1 << (5 - 1)) - 1) + 1,                                // eMin
     ((((1 << (5 - 1)) - 1) + 127) << 23) | 0x7FE000,          // maxNormal
     ((-((1 << (5 - 1)) - 1) + 1) + 127) << 23,                // minNormal
-    static_cast<uint32>((((1 << (5 - 1)) - 1) - 127) << 23),  // biasDiff
+    (((1U << (5 - 1)) - 1) - 127) << 23,                      // biasDiff
     (23 - 10),                                                // fracBitsDiff
 };
 
@@ -86,7 +86,7 @@ static constexpr NBitFloatInfo Float11Info =
     -((1 << (5 - 1)) - 1)+1,                                  // eMin
     ((((1 << (5 - 1)) - 1) + 127) << 23) | 0x7E000,           // maxNormal
     ((-((1 << (5 - 1)) - 1) + 1) + 127) << 23,                // minNormal
-    static_cast<uint32>((((1 << (5 - 1)) - 1) - 127) << 23),  // biasDiff
+    (((1U << (5 - 1)) - 1) - 127) << 23,                      // biasDiff
     23 - 6,                                                   // fracBitsDiff
 };
 
@@ -104,7 +104,7 @@ static constexpr NBitFloatInfo Float10Info =
     -((1 << (5 - 1)) - 1)+1,                                  // eMin
     ((((1 << (5 - 1)) - 1) + 127) << 23) | 0x7C000,           // maxNormal
     ((-((1 << (5 - 1)) - 1) + 1) + 127) << 23,                // minNormal
-    static_cast<uint32>((((1 << (5 - 1)) - 1) - 127) << 23),  // biasDiff
+    (((1U << (5 - 1)) - 1) - 127) << 23,                      // biasDiff
     23 - 5,                                                   // fracBitsDiff
 };
